Create hook to scroll to specified notifications
